requirements.txt
streamlit
google-generativeai
openai
python-dotenv
requests
